Redfactor to password class
