+++ app/models/user.rb
@@ -6,12 +6,6 @@ class User < ActiveRecord::Base
   validates_presence_of :username, :forename, :surname, :email, :password
   validates_uniqueness_of :username, :email
 
-  def self.check_password(username, password)
-    stored_password = User.where(username: username).pluck(:password).first
-    return false if stored_password.nil? || BCrypt::Password.new(stored_password) != password
-    true
-  end
-
   def self.assign_user(username)
     @@logged_in_name = username
   end
